Struct identity_did::document::document::Document

pub struct Document<T = Object, U = Object, V = Object> {
    pub(crate) id: DID,
    pub(crate) controller: Option<DID>,
    pub(crate) also_known_as: Vec<Url>,
    pub(crate) verification_method: OrderedSet<DIDKey<Method<U>>>,
    pub(crate) authentication: OrderedSet<DIDKey<MethodRef<U>>>,
    pub(crate) assertion_method: OrderedSet<DIDKey<MethodRef<U>>>,
    pub(crate) key_agreement: OrderedSet<DIDKey<MethodRef<U>>>,
    pub(crate) capability_delegation: OrderedSet<DIDKey<MethodRef<U>>>,
    pub(crate) capability_invocation: OrderedSet<DIDKey<MethodRef<U>>>,
    pub(crate) service: OrderedSet<DIDKey<Service<V>>>,
    pub(crate) properties: T,
}

A DID Document

Fields
id: DIDcontroller: Option<DID>also_known_as: Vec<Url>verification_method: OrderedSet<DIDKey<Method<U>>>authentication: OrderedSet<DIDKey<MethodRef<U>>>assertion_method: OrderedSet<DIDKey<MethodRef<U>>>key_agreement: OrderedSet<DIDKey<MethodRef<U>>>capability_delegation: OrderedSet<DIDKey<MethodRef<U>>>capability_invocation: OrderedSet<DIDKey<MethodRef<U>>>service: OrderedSet<DIDKey<Service<V>>>properties: T
Implementations

impl<T, U, V> Document<T, U, V>

pub fn builder(properties: T) -> DocumentBuilder<T, U, V>

Creates a DocumentBuilder to configure a new Document.

This is the same as DocumentBuilder::new().

pub fn from_builder(builder: DocumentBuilder<T, U, V>) -> Result<Self>

Returns a new Document based on the DocumentBuilder configuration.

pub fn id(&self) -> &DID

Returns a reference to the Document id.

pub fn id_mut(&mut self) -> &mut DID

Returns a mutable reference to the Document id.

pub fn controller(&self) -> Option<&DID>

Returns a reference to the Document controller.

pub fn controller_mut(&mut self) -> Option<&mut DID>

Returns a mutable reference to the Document controller.

pub fn also_known_as(&self) -> &[Url]

Returns a reference to the Document alsoKnownAs set.

pub fn also_known_as_mut(&mut self) -> &mut Vec<Url>

Returns a mutable reference to the Document alsoKnownAs set.

pub fn verification_method(&self) -> &OrderedSet<DIDKey<Method<U>>>

Returns a reference to the Document verificationMethod set.

pub fn verification_method_mut(&mut self) -> &mut OrderedSet<DIDKey<Method<U>>>

Returns a mutable reference to the Document verificationMethod set.

pub fn authentication(&self) -> &OrderedSet<DIDKey<MethodRef<U>>>

Returns a reference to the Document authentication set.

pub fn authentication_mut(&mut self) -> &mut OrderedSet<DIDKey<MethodRef<U>>>

Returns a mutable reference to the Document authentication set.

pub fn assertion_method(&self) -> &OrderedSet<DIDKey<MethodRef<U>>>

Returns a reference to the Document assertionMethod set.

pub fn assertion_method_mut(&mut self) -> &mut OrderedSet<DIDKey<MethodRef<U>>>

Returns a mutable reference to the Document assertionMethod set.

pub fn key_agreement(&self) -> &OrderedSet<DIDKey<MethodRef<U>>>

Returns a reference to the Document keyAgreement set.

pub fn key_agreement_mut(&mut self) -> &mut OrderedSet<DIDKey<MethodRef<U>>>

Returns a mutable reference to the Document keyAgreement set.

pub fn capability_delegation(&self) -> &OrderedSet<DIDKey<MethodRef<U>>>

Returns a reference to the Document capabilityDelegation set.

pub fn capability_delegation_mut(
&mut self
) -> &mut OrderedSet<DIDKey<MethodRef<U>>>

Returns a mutable reference to the Document capabilityDelegation set.

pub fn capability_invocation(&self) -> &OrderedSet<DIDKey<MethodRef<U>>>

Returns a reference to the Document capabilityInvocation set.

pub fn capability_invocation_mut(
&mut self
) -> &mut OrderedSet<DIDKey<MethodRef<U>>>

Returns a mutable reference to the Document capabilityInvocation set.

pub fn service(&self) -> &OrderedSet<DIDKey<Service<V>>>

Returns a reference to the Document service set.

pub fn service_mut(&mut self) -> &mut OrderedSet<DIDKey<Service<V>>>

Returns a mutable reference to the Document service set.

pub fn properties(&self) -> &T

Returns a reference to the custom Document properties.

pub fn properties_mut(&mut self) -> &mut T

Returns a mutable reference to the custom Document properties.

pub fn map<A, F>(self, f: F) -> Document<A, U, V> where
F: FnOnce(T) -> A,

Maps Document<T> to Document<U> by applying a function to the custom properties.

pub fn try_map<A, F, E>(self, f: F) -> Result<Document<A, U, V>, E> where
F: FnOnce(T) -> Result<A, E>,

A fallible version of Document::map(..).

Errors

try_map can fail if the provided function fails.

pub fn insert_method(&mut self, scope: MethodScope, method: Method<U>) -> bool

Adds a new Method<U> to the Document.

pub fn remove_method(&mut self, did: &DID)

Removes all references to the specified Method<U>.

pub fn methods(&self) -> impl Iterator<Item = &Method<U>>

Returns an iterator over all verification methods in the DID Document.

pub fn resolve<'query, Q>(&self, query: Q) -> Option<&Method<U>> where
Q: Into<MethodQuery<'query>>,

Returns the first verification method with an id property matching the provided query.

pub fn try_resolve<'query, Q>(&self, query: Q) -> Result<&Method<U>> where
Q: Into<MethodQuery<'query>>,

Returns the first verification method with an id property matching the provided query.

Errors

Fails if no matching verification Method is found.

pub fn resolve_mut<'query, Q>(&mut self, query: Q) -> Option<&mut Method<U>> where
Q: Into<MethodQuery<'query>>,

Returns a mutable reference to the first verification method with an id property matching the provided query.

pub fn try_resolve_mut<'query, Q>(&mut self, query: Q) -> Result<&mut Method<U>> where
Q: Into<MethodQuery<'query>>,

Returns a mutable reference to the first verification method with an id property matching the provided query.

Errors

Fails if no matching verification Method is found.

fn resolve_method(&self, query: MethodQuery<'_>) -> Option<&Method<U>>

fn resolve_method_mut(
&mut self,
query: MethodQuery<'_>
) -> Option<&mut Method<U>>

impl<T, U, V> Document<T, U, V>

pub fn into_verifiable(self) -> Document<Properties<T>, U, V>

pub fn into_verifiable2(self, proof: Signature) -> Document<Properties<T>, U, V>

impl<T, U, V> Document<Properties<T>, U, V>

pub fn proof(&self) -> Option<&Signature>

pub fn proof_mut(&mut self) -> Option<&mut Signature>

pub fn set_proof(&mut self, signature: Signature)

impl<T, U, V> Document<Properties<T>, U, V> where
T: Serialize,
U: Serialize,
V: Serialize,

pub fn sign_this<'query, Q>(&mut self, query: Q, secret: &[u8]) -> Result<()> where
Q: Into<MethodQuery<'query>>,

pub fn verify_this(&self) -> Result<()>

impl<T, U, V> Document<T, U, V>

pub fn signer<'base>(
&'base self,
secret: &'base SecretKey
) -> DocumentSigner<'base, '_, '_, T, U, V>

Creates a new DocumentSigner that can be used to create digital signatures from verification methods in this DID Document.

pub fn verifier(&self) -> DocumentVerifier<'_, T, U, V>

Creates a new DocumentVerifier that can be used to verify signatures created with this DID Document.