Struct identity_did::document::Document

pub struct Document<T = (), U = (), V = ()> {
    pub(crate) id: DID,
    pub(crate) controller: Option<DID>,
    pub(crate) also_known_as: Vec<Url>,
    pub(crate) verification_method: OrderedSet<DIDKey<Method<U>>>,
    pub(crate) authentication: OrderedSet<DIDKey<MethodRef<U>>>,
    pub(crate) assertion_method: OrderedSet<DIDKey<MethodRef<U>>>,
    pub(crate) key_agreement: OrderedSet<DIDKey<MethodRef<U>>>,
    pub(crate) capability_delegation: OrderedSet<DIDKey<MethodRef<U>>>,
    pub(crate) capability_invocation: OrderedSet<DIDKey<MethodRef<U>>>,
    pub(crate) service: OrderedSet<DIDKey<Service<V>>>,
    pub(crate) properties: T,
}

A DID Document

Fields
id: DIDcontroller: Option<DID>also_known_as: Vec<Url>verification_method: OrderedSet<DIDKey<Method<U>>>authentication: OrderedSet<DIDKey<MethodRef<U>>>assertion_method: OrderedSet<DIDKey<MethodRef<U>>>key_agreement: OrderedSet<DIDKey<MethodRef<U>>>capability_delegation: OrderedSet<DIDKey<MethodRef<U>>>capability_invocation: OrderedSet<DIDKey<MethodRef<U>>>service: OrderedSet<DIDKey<Service<V>>>properties: T
Implementations

impl<T, U, V> Document<T, U, V>

pub fn builder(properties: T) -> DocumentBuilder<T, U, V>

Creates a DocumentBuilder to configure a new Document.

This is the same as DocumentBuilder::new().

pub fn from_builder(builder: DocumentBuilder<T, U, V>) -> Result<Self>

Returns a new Document based on the DocumentBuilder configuration.

pub fn id(&self) -> &DID

Returns a reference to the Document id.

pub fn id_mut(&mut self) -> &mut DID

Returns a mutable reference to the Document id.

pub fn controller(&self) -> Option<&DID>

Returns a reference to the Document controller.

pub fn controller_mut(&mut self) -> Option<&mut DID>

Returns a mutable reference to the Document controller.

pub fn also_known_as(&self) -> &[Url]

Returns a reference to the Document alsoKnownAs set.

pub fn also_known_as_mut(&mut self) -> &mut Vec<Url>

Returns a mutable reference to the Document alsoKnownAs set.

pub fn verification_method(&self) -> &OrderedSet<DIDKey<Method<U>>>

Returns a reference to the Document verificationMethod set.

pub fn verification_method_mut(&mut self) -> &mut OrderedSet<DIDKey<Method<U>>>

Returns a mutable reference to the Document verificationMethod set.

pub fn authentication(&self) -> &OrderedSet<DIDKey<MethodRef<U>>>

Returns a reference to the Document authentication set.

pub fn authentication_mut(&mut self) -> &mut OrderedSet<DIDKey<MethodRef<U>>>

Returns a mutable reference to the Document authentication set.

pub fn assertion_method(&self) -> &OrderedSet<DIDKey<MethodRef<U>>>

Returns a reference to the Document assertionMethod set.

pub fn assertion_method_mut(&mut self) -> &mut OrderedSet<DIDKey<MethodRef<U>>>

Returns a mutable reference to the Document assertionMethod set.

pub fn key_agreement(&self) -> &OrderedSet<DIDKey<MethodRef<U>>>

Returns a reference to the Document keyAgreement set.

pub fn key_agreement_mut(&mut self) -> &mut OrderedSet<DIDKey<MethodRef<U>>>

Returns a mutable reference to the Document keyAgreement set.

pub fn capability_delegation(&self) -> &OrderedSet<DIDKey<MethodRef<U>>>

Returns a reference to the Document capabilityDelegation set.

pub fn capability_delegation_mut(
&mut self
) -> &mut OrderedSet<DIDKey<MethodRef<U>>>

Returns a mutable reference to the Document capabilityDelegation set.

pub fn capability_invocation(&self) -> &OrderedSet<DIDKey<MethodRef<U>>>

Returns a reference to the Document capabilityInvocation set.

pub fn capability_invocation_mut(
&mut self
) -> &mut OrderedSet<DIDKey<MethodRef<U>>>

Returns a mutable reference to the Document capabilityInvocation set.

pub fn service(&self) -> &OrderedSet<DIDKey<Service<V>>>

Returns a reference to the Document service set.

pub fn service_mut(&mut self) -> &mut OrderedSet<DIDKey<Service<V>>>

Returns a mutable reference to the Document service set.

pub fn properties(&self) -> &T

Returns a reference to the custom Document properties.

pub fn properties_mut(&mut self) -> &mut T

Returns a mutable reference to the custom Document properties.

pub fn map<A, F>(self, f: F) -> Document<A, U, V> where
F: FnOnce(T) -> A,

Maps Document<T> to Document<U> by applying a function to the custom properties.

pub fn try_map<A, F, E>(self, f: F) -> Result<Document<A, U, V>, E> where
F: FnOnce(T) -> Result<A, E>,

A fallible version of Document::map(..).

Errors

try_map can fail if the provided function fails.

pub fn resolve<'a, Q>(&self, query: Q) -> Option<MethodWrap<'_, U>> where
Q: Into<MethodQuery<'a>>,

Finds and returns the first verification Method matching the provided MethodQuery.

pub fn try_resolve<'a, Q>(&self, query: Q) -> Result<MethodWrap<'_, U>> where
Q: Into<MethodQuery<'a>>,

Finds and returns the first verification Method matching the provided MethodQuery.

Errors

Fails if no matching verification Method is found.

pub fn resolve_bytes<'a, Q>(&self, query: Q) -> Option<Vec<u8>> where
Q: Into<MethodQuery<'a>>,

pub fn try_resolve_bytes<'a, Q>(&self, query: Q) -> Result<Vec<u8>> where
Q: Into<MethodQuery<'a>>,

pub fn resolve_options<'a, Q>(&self, query: Q) -> Result<SignatureOptions> where
Q: Into<MethodQuery<'a>>,

fn resolve_method<'a>(
&self,
query: MethodQuery<'a>
) -> Option<MethodWrap<'_, U>>

fn resolve_verification_method(
&self,
query: MethodQuery<'_>
) -> Option<MethodWrap<'_, U>>

impl<T, U, V> Document<T, U, V>

pub fn into_verifiable(self) -> Document<Properties<T>, U, V>

pub fn into_verifiable2(self, proof: Signature) -> Document<Properties<T>, U, V>

impl<T, U, V> Document<Properties<T>, U, V>

pub fn proof(&self) -> Option<&Signature>

pub fn proof_mut(&mut self) -> Option<&mut Signature>

pub fn set_proof(&mut self, signature: Signature)