Module identity_did::verification

Modules

builder

method

method_data

method_ident

method_query

method_ref

method_scope

method_type

method_wrap

Structs

Method
A DID Document Verification Method

MethodBuilder
A MethodBuilder is used to generate a customized Method.

MethodQuery
Specifies the conditions of a DID document method resolution query.

MethodWrap
A queried Method with additional metadata about the query resolution.

Enums

MethodData
Supported verification method data formats.

MethodIdent
Index or identifier used to identify the target verification method of a MethodQuery.

MethodRef
A reference to a verification method, either a DID or embedded Method.

MethodScope
Verification method group used to refine the scope of a method query.

MethodType
Supported verification method types.